feat(core): Introduce identity-substrate module and move substrate-owned files into it (no-changelog)
Splits packages/cli/src/modules/token-exchange into a substrate and a
consumer (IAM-1181). This first step moves the entities, repositories,
and the pure-utility services (JWKS resolution, replay/jti store and
cleanup, bearer-token extraction) that carry no main-only behavior of
their own into a new identity-substrate module. The module currently
declares only entities() — no init() yet — so this is a relocation
with no behavior change: token-exchange still does everything it did
before, just importing the moved pieces cross-module.

Table names are unchanged; no DB migration is needed since entity
registration in ModuleRegistry.loadModules() is unconditional across
modules, independent of instanceTypes/licenseFlag.
