+++ packages/@n8n/backend-common/src/modules/__tests__/module-registry.test.ts
@@ -46,6 +46,7 @@ describe('eligibleModules', () => {
 			'source-control',
 			'dynamic-credentials',
 			'chat-hub',
+			'identity-substrate',
 			'token-exchange',
 			'sso-oidc',
 			'sso-saml',
@@ -81,6 +82,7 @@ describe('eligibleModules', () => {
 			'source-control',
 			'dynamic-credentials',
 			'chat-hub',
+			'identity-substrate',
 			'token-exchange',
 			'sso-oidc',
 			'sso-saml',
