+++ packages/@n8n/backend-common/src/modules/modules.config.ts
@@ -28,6 +28,7 @@ export const MODULE_NAMES = [
 	'instance-ai',
 	'mcp-registry',
 	'otel',
+	'identity-substrate',
 	'token-exchange',
 	'instance-version-history',
 	'encryption-key-manager',
