+++ packages/cli/src/modules/token-exchange/controllers/__tests__/trusted-key-source.controller.test.ts
@@ -3,7 +3,8 @@ import type { Logger } from '@n8n/backend-common';
 import type { AuthenticatedRequest } from '@n8n/db';
 import { mock } from 'vitest-mock-extended';
 
-import type { TrustedKeySourceEntity } from '../../database/entities/trusted-key-source.entity';
+import type { TrustedKeySourceEntity } from '@/modules/identity-substrate/database/entities/trusted-key-source.entity';
+
 import type { TrustedKeyService } from '../../services/trusted-key.service';
 import { TrustedKeySourceController } from '../trusted-key-source.controller';
