+++ packages/cli/test/integration/identity-substrate/token-exchange-jti.repository.integration.test.ts
@@ -1,13 +1,13 @@
 import { testDb, testModules } from '@n8n/backend-test-utils';
 import { Container } from '@n8n/di';
 
-import { TokenExchangeJtiRepository } from '@/modules/token-exchange/database/repositories/token-exchange-jti.repository';
+import { TokenExchangeJtiRepository } from '@/modules/identity-substrate/database/repositories/token-exchange-jti.repository';
 
 describe('TokenExchangeJtiRepository', () => {
 	let repository: TokenExchangeJtiRepository;
 
 	beforeAll(async () => {
-		await testModules.loadModules(['token-exchange']);
+		await testModules.loadModules(['identity-substrate']);
 		await testDb.init();
 		repository = Container.get(TokenExchangeJtiRepository);
 	});
