+++ packages/cli/test/integration/token-exchange/trusted-key.service.integration.test.ts
@@ -3,10 +3,10 @@ import { Container } from '@n8n/di';
 import { InstanceSettings } from 'n8n-core';
 import type { KeyObject } from 'node:crypto';
 
-import type { TrustedKeySourceEntity } from '@/modules/token-exchange/database/entities/trusted-key-source.entity';
-import { TrustedKeyEntity } from '@/modules/token-exchange/database/entities/trusted-key.entity';
-import { TrustedKeySourceRepository } from '@/modules/token-exchange/database/repositories/trusted-key-source.repository';
-import { TrustedKeyRepository } from '@/modules/token-exchange/database/repositories/trusted-key.repository';
+import type { TrustedKeySourceEntity } from '@/modules/identity-substrate/database/entities/trusted-key-source.entity';
+import { TrustedKeyEntity } from '@/modules/identity-substrate/database/entities/trusted-key.entity';
+import { TrustedKeySourceRepository } from '@/modules/identity-substrate/database/repositories/trusted-key-source.repository';
+import { TrustedKeyRepository } from '@/modules/identity-substrate/database/repositories/trusted-key.repository';
 import { TrustedKeyService } from '@/modules/token-exchange/services/trusted-key.service';
 import { TokenExchangeConfig } from '@/modules/token-exchange/token-exchange.config';
 import type { TrustedKeyData } from '@/modules/token-exchange/token-exchange.schemas';
@@ -109,7 +109,7 @@ let keyRepo: TrustedKeyRepository;
 let instanceSettings: InstanceSettings;
 
 beforeAll(async () => {
-	await testModules.loadModules(['token-exchange']);
+	await testModules.loadModules(['identity-substrate', 'token-exchange']);
 	await testDb.init();
 
 	instanceSettings = Container.get(InstanceSettings);
